fix: retrieve subjects via User::tutorProfile relationship
In commit 30f507c we defined the M-M relationship between Tutor and
Subject instead of between User and Subject. Hence we should have
updated the TutorIndex Livewire page component to reflect this
change.

We've also renamed $tutors to $users since we are fetching the tutors
as User records rather than as Tutor records.